service feat: add user adress apis

+++ packages/service/src/modules/users/users.service.ts
@@ -14,19 +14,21 @@ import { UsersChangePasswordReqDto } from './dtos/users.change-password.req.dto'
 import { UserFiltered } from 'src/shared/interfaces/user.filtered.res.interface';
 import { PasswordUpdateDto } from 'src/common/mongodb/usersdb/dtos/password.update.dto';
 import { UserUpdateDto } from 'src/common/mongodb/usersdb/dtos/user.update.dto';
-import { UsersPermsCollService } from 'src/common/mongodb/usersdb/services/users-permissions.collection.service';
-import { filterUserRes } from 'src/shared/utils/filterUserRes';
-import { UserPermsRes } from 'src/common/mongodb/usersdb/interfaces/user-permissions.res.interface';
+import { UserAddrsRes } from 'src/common/mongodb/usersdb/interfaces/user-addresses.res.interface';
+import { UsersAddrsCollService } from 'src/common/mongodb/usersdb/services/users-addresses.collection/users-addresses.collection.service';
+import { UserAddrDto } from './dtos/user.address.dto';
+import { UserAddrUpdateReqDto } from './dtos/user-address.update.req.dto';
+import { cvtToObjectId } from 'src/shared/utils/cvtToObjectId';
 
 @Injectable()
 export class UsersService {
     constructor(
         private readonly usersCollService: UsersCollService,
-        private readonly usersPermsCollService: UsersPermsCollService,
+        private readonly usersAddrsCollService: UsersAddrsCollService,
     ) {}
 
     async getProfile(userId: Types.ObjectId): Promise<UserFiltered> {
-        return await this.usersCollService.getWithPerms(userId);
+        return await this.usersCollService.getUserFiltered(userId);
     }
 
     async updateProfile(
@@ -55,38 +57,33 @@ export class UsersService {
             throw new InternalServerErrorException();
         }
 
-        return this.usersCollService.getWithPerms(userId);
+        return this.usersCollService.getUserFiltered(userId);
     }
 
     async deleteProfile(
         userId: Types.ObjectId,
         password: string,
     ): Promise<UserFiltered> {
-        await this.comparePassword(
-            userId,
-            password,
-            'The password is incorrect',
-        );
+        await comparePassword(userId, password, 'The password is incorrect');
+
+        const user = await this.usersCollService.getUserFiltered(userId);
+        if (!user) {
+            throw new InternalServerErrorException();
+        }
 
         const deletedUser = await this.usersCollService.delete(userId);
         if (!deletedUser) {
             throw new InternalServerErrorException();
         }
 
-        let deletedPermission: UserPermsRes;
-        if (deletedUser.permissions) {
-            deletedPermission =
-                await this.usersPermsCollService.deleteByUserId(userId);
-        }
-
-        return filterUserRes(deletedUser, deletedPermission);
+        return user;
     }
 
     async changePassword(
         userId: Types.ObjectId,
         usersChangePasswordReqDto: UsersChangePasswordReqDto,
     ): Promise<UserFiltered> {
-        await this.comparePassword(
+        await comparePassword(
             userId,
             usersChangePasswordReqDto.password,
             'The old password is incorrect',
@@ -111,22 +108,114 @@ export class UsersService {
             throw new InternalServerErrorException();
         }
 
-        return await this.usersCollService.getWithPerms(userId);
+        return await this.usersCollService.getUserFiltered(userId);
     }
 
-    private async comparePassword(
-        userId: Types.ObjectId,
-        password: string,
-        errMsg: string,
-    ): Promise<void> {
-        const user = await this.usersCollService.findById(userId);
-        if (!user) {
-            throw new NotFoundException();
+    async getAddresses(userId: Types.ObjectId): Promise<UserAddrsRes[]> {
+        const addrs = await this.usersCollService.getAddresses(userId);
+        if (!addrs) {
+            throw new InternalServerErrorException();
         }
 
-        const isMatch = await bcrypt.compare(password, user.password);
-        if (!isMatch) {
-            throw new HttpException(errMsg, HttpStatus.UNAUTHORIZED);
+        return addrs;
+    }
+
+    async addAddress(
+        userId: Types.ObjectId,
+        userAddrDto: UserAddrDto,
+    ): Promise<UserAddrsRes> {
+        const user = await this.usersCollService.findById(userId);
+        if (user.addresses.length >= 5) {
+            throw new HttpException(
+                'The user can have at most 5 addresses',
+                HttpStatus.BAD_REQUEST,
+            );
         }
+
+        const addr = await this.usersAddrsCollService.saveNew(userAddrDto);
+        if (!addr) {
+            throw new InternalServerErrorException();
+        }
+
+        const newUser = await this.usersCollService.addAddress(
+            userId,
+            addr._id,
+        );
+        if (!newUser) {
+            throw new InternalServerErrorException();
+        }
+
+        return addr;
+    }
+
+    async updateAddress(
+        userId: Types.ObjectId,
+        addrId: string,
+        userAddrUpdateReqDto: UserAddrUpdateReqDto,
+    ): Promise<UserAddrsRes> {
+        const addrObjId = cvtToObjectId(addrId, 'addrId');
+
+        const user = await this.usersCollService.findById(userId);
+        if (!user) {
+            throw new InternalServerErrorException();
+        }
+        if (!user.addresses.find((id) => String(id) === String(addrObjId))) {
+            throw new HttpException(
+                'The user does not own this address',
+                HttpStatus.BAD_REQUEST,
+            );
+        }
+
+        return this.usersAddrsCollService.updateById(
+            addrObjId,
+            userAddrUpdateReqDto,
+        );
+    }
+
+    async removeAddress(userId: Types.ObjectId, addrId: string) {
+        const addrObjId = cvtToObjectId(addrId, 'addrId');
+
+        const user = await this.usersCollService.findById(userId);
+        if (!user) {
+            throw new InternalServerErrorException();
+        }
+        if (!user.addresses.find((id) => String(id) === String(addrObjId))) {
+            throw new HttpException(
+                'The user does not own this address',
+                HttpStatus.BAD_REQUEST,
+            );
+        }
+
+        const deletedAddr =
+            await this.usersAddrsCollService.removeById(addrObjId);
+        if (!deletedAddr) {
+            throw new InternalServerErrorException();
+        }
+
+        const newUser = await this.usersCollService.removeAddress(
+            userId,
+            deletedAddr._id,
+        );
+        if (!newUser) {
+            throw new InternalServerErrorException();
+        }
+
+        return deletedAddr;
+    }
+}
+
+async function comparePassword(
+    userId: Types.ObjectId,
+    password: string,
+    errMsg: string,
+): Promise<void> {
+    const user = await this.usersCollService.findById(userId);
+    if (!user) {
+        throw new NotFoundException();
+    }
+
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) {
+        throw new HttpException(errMsg, HttpStatus.UNAUTHORIZED);
     }
 }
